backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
sqlalchemy==2.0.35
psycopg2-binary>=2.9
python-dotenv==1.0.1
